+++ digest/src/core_api.rs
@@ -8,10 +8,7 @@ use crate::InvalidOutputSize;
 pub use crypto_common::{AlgorithmName, Block, BlockSizeUser, OutputSizeUser, Reset};
 
 use block_buffer::{BlockBuffer, BufferKind};
-use crypto_common::{
-    typenum::{IsLess, Le, NonZero, U256},
-    Output,
-};
+use crypto_common::Output;
 
 mod ct_variable;
 mod rt_variable;
@@ -40,22 +37,14 @@ pub trait BufferKindUser: BlockSizeUser {
 }
 
 /// Core trait for hash functions with fixed output size.
-pub trait FixedOutputCore: UpdateCore + BufferKindUser + OutputSizeUser
-where
-    Self::BlockSize: IsLess<U256>,
-    Le<Self::BlockSize, U256>: NonZero,
-{
+pub trait FixedOutputCore: UpdateCore + BufferKindUser + OutputSizeUser {
     /// Finalize state using remaining data stored in the provided block buffer,
     /// write result into provided array and leave `self` in a dirty state.
     fn finalize_fixed_core(&mut self, buffer: &mut Buffer<Self>, out: &mut Output<Self>);
 }
 
 /// Core trait for hash functions with extendable (XOF) output size.
-pub trait ExtendableOutputCore: UpdateCore + BufferKindUser
-where
-    Self::BlockSize: IsLess<U256>,
-    Le<Self::BlockSize, U256>: NonZero,
-{
+pub trait ExtendableOutputCore: UpdateCore + BufferKindUser {
     /// XOF reader core state.
     type ReaderCore: XofReaderCore;
 
@@ -81,11 +70,7 @@ pub trait XofReaderCore: BlockSizeUser {
 /// [`finalize_variable_core`]: VariableOutputCore::finalize_variable_core
 /// [`new`]: VariableOutputCore::new
 /// [`TRUNC_SIDE`]: VariableOutputCore::TRUNC_SIDE
-pub trait VariableOutputCore: UpdateCore + OutputSizeUser + BufferKindUser + Sized
-where
-    Self::BlockSize: IsLess<U256>,
-    Le<Self::BlockSize, U256>: NonZero,
-{
+pub trait VariableOutputCore: UpdateCore + OutputSizeUser + BufferKindUser + Sized {
     /// Side which should be used in a truncated result.
     const TRUNC_SIDE: TruncSide;