+++ digest/src/core_api/xof_reader.rs
@@ -1,8 +1,7 @@
 use super::{AlgorithmName, XofReaderCore};
 use crate::XofReader;
-use block_buffer::EagerBuffer;
+use block_buffer::ReadBuffer;
 use core::fmt;
-use crypto_common::typenum::{IsLess, Le, NonZero, U256};
 
 /// Wrapper around [`XofReaderCore`] implementations.
 ///
@@ -11,18 +10,14 @@ use crypto_common::typenum::{IsLess, Le, NonZero, U256};
 pub struct XofReaderCoreWrapper<T>
 where
     T: XofReaderCore,
-    T::BlockSize: IsLess<U256>,
-    Le<T::BlockSize, U256>: NonZero,
 {
     pub(super) core: T,
-    pub(super) buffer: EagerBuffer<T::BlockSize>,
+    pub(super) buffer: ReadBuffer<T::BlockSize>,
 }
 
 impl<T> fmt::Debug for XofReaderCoreWrapper<T>
 where
     T: XofReaderCore + AlgorithmName,
-    T::BlockSize: IsLess<U256>,
-    Le<T::BlockSize, U256>: NonZero,
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
         T::write_alg_name(f)?;
@@ -33,17 +28,11 @@ where
 impl<T> XofReader for XofReaderCoreWrapper<T>
 where
     T: XofReaderCore,
-    T::BlockSize: IsLess<U256>,
-    Le<T::BlockSize, U256>: NonZero,
 {
     #[inline]
     fn read(&mut self, buffer: &mut [u8]) {
         let Self { core, buffer: buf } = self;
-        buf.set_data(buffer, |blocks| {
-            for block in blocks {
-                *block = core.read_block();
-            }
-        });
+        buf.read(buffer, |block| *block = core.read_block());
     }
 }
 
@@ -52,8 +41,6 @@ where
 impl<T> std::io::Read for XofReaderCoreWrapper<T>
 where
     T: XofReaderCore,
-    T::BlockSize: IsLess<U256>,
-    Le<T::BlockSize, U256>: NonZero,
 {
     #[inline]
     fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {